pyproject.toml
[tool.poetry]
name = "CDN-Jump"
version = "0.1.0"
description = "Herramienta en Python para recopilar y analizar información de dominios, integrando consultas DNS, APIs externas y validaciones."
authors = ["goalnefesh <goalnefesh@protonmail.com>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "cdn_jump" },
    { include = "config" }
]

[tool.poetry.dependencies]
python = ">=3.9, <3.14"
pip = "^25.0.1"
beautifulsoup4 = "^4.13.3"
dnspython = "^2.7.0"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
censys = "^2.2.16"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
flake8 = "^5.0.4"
black = "^23.3.0"

[build-system]
requires = ["poetry-core>=1.4.0"]
build-backend = "poetry.core.masonry.api"